requirements.txt
termcolor
pillow
numpy
PySide2